requirements.txt
Pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
balethon
requests
opencv-python==4.8.1.78  
numpy==1.26.0           
scikit-learn==1.3.1 
